add types for Login, Messages
