+++ src/App.js
@@ -4,7 +4,7 @@ import Nav from './compo/Nav/Nav'
 import {Route, Routes} from '../node_modules/react-router-dom/index'
 import {Navigate} from 'react-router-dom'
 import ScrollToTop from 'react-scroll-to-top'
-import Login from './compo/Login/Login'
+import Login from './compo/Login/Login.tsx'
 import {useDispatch, useSelector} from 'react-redux'
 import {lazy, Suspense, useEffect} from 'react'
 import {getInitialized} from './redux/selectors/appSelector'
